# Fix CookieContainer memory leak