Cookies are correctly removed based on `Capacity` and `PerDomainCapacity` from `CookieCollection` in each `PathList`, but we forget to clean up the `DomainTable` even when there is no/empty `PathList` associated with a `domainKey`. This can result in memory leak, particularly with these two cases:
1. Add cookies for many distinct domains.
2. Add cookies for the same domain with different paths.

Fix #33712.

Update, post `BenchmarkDotNet` result - the improvement is huge:

Before:

```
// * Summary *

BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17763, VM=Hyper-V
Intel Xeon CPU E5-2673 v4 2.30GHz, 1 CPU, 4 logical and 2 physical cores
.NET Core SDK=2.1.500
  [Host]     : .NET Core 2.1.6 (CoreCLR 4.6.27019.06, CoreFX 4.6.27019.05), 64bit RyuJIT
  Job-ZPUWDY : .NET Core d4939a36-e31f-4770-967a-35a619c12264 (CoreCLR 4.6.27304.02, CoreFX 4.7.19.5701), 64bit RyuJIT

Runtime=Core  Toolchain=CoreRun

                        Method |     Mean |    Error |   StdDev |       Gen 0 |       Gen 1 |     Gen 2 | Allocated |
------------------------------ |---------:|---------:|---------:|------------:|------------:|----------:|----------:|
      AddCookiesFromSameDomain | 38.210 s | 0.6532 s | 0.5791 s | 932000.0000 | 137000.0000 | 2000.0000 |   5.47 GB |
 AddCookiesFromDifferentDomain |  4.610 s | 0.0895 s | 0.0747 s | 189000.0000 |  12000.0000 | 2000.0000 |   1.13 GB |
```

After:

```
// * Summary *

BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17763, VM=Hyper-V
Intel Xeon CPU E5-2673 v4 2.30GHz, 1 CPU, 4 logical and 2 physical cores
.NET Core SDK=2.1.500
  [Host]     : .NET Core 2.1.6 (CoreCLR 4.6.27019.06, CoreFX 4.6.27019.05), 64bit RyuJIT
  Job-DGWAED : .NET Core 90a509e8-00cd-4c3e-bac4-d99eb4e58bfa (CoreCLR 4.6.27304.02, CoreFX 4.7.19.5701), 64bit RyuJIT

Runtime=Core  Toolchain=CoreRun

                        Method |     Mean |     Error |    StdDev |     Gen 0 | Allocated |
------------------------------ |---------:|----------:|----------:|----------:|----------:|
      AddCookiesFromSameDomain | 69.75 ms | 1.5412 ms | 4.1667 ms |         - |  15.91 MB |
 AddCookiesFromDifferentDomain | 44.38 ms | 0.7127 ms | 0.6666 ms | 3083.3333 |  20.23 MB |
```

Here is the test code if you are interested:

```c#

        public void AddCookies_OldCookiesRemoved(bool isFromSameDomain)
        {
            const int Capacity = 10;
            const int TotalCookieCount = 5000;
            var cookieContainer = new CookieContainer(Capacity);
            Cookie cookie;
            for (int i = 0; i < TotalCookieCount; i++)
            {
                if (isFromSameDomain)
                {
                    cookie = new Cookie("name1", "value1", $"/{i}", "test.com");
                }
                else
                {
                    cookie = new Cookie("name1", "value1", "/", $"test{i}.com");
                }
                cookieContainer.Add(cookie);
            }
        }

        [Benchmark] public void AddCookiesFromSameDomain() => AddCookies_OldCookiesRemoved(true);
        [Benchmark] public void AddCookiesFromDifferentDomain() => AddCookies_OldCookiesRemoved(false);

```