@caesar1995 2019-01-10
@dotnet-bot test UWP CoreCLR x64 Debug Build please
@dotnet-bot test outerloop linux x64 debug build
@dotnet-bot test outerloop windows x64 debug build
@dotnet-bot test outerloop OSX x64 debug build